Revert "frontend: fix snr value Sundtek DVB-T (III)"
This reverts commit 963266903c6c0b66f4012a2a9ae18d21746e9242.

When either signalquality or signalqualitydb are nonzero, the driver
implements the DTV_STAT_CNR property.
And we have to trust the driver to report the correct values, no need to
make any exceptions.

Perhaps this commit was meant to force snr percentage to dB calculation
for Sundtek, but in that case a different, more generic fix should be
made to calculate dB values based on the FE_SCALE_RELATIVE value
reported by the DTV_STAT_CNR property.