Support ac3 and mka audio formats (#598)
ac3 is being used for dubbed movies.
mka is the audio vesion of mkv.
Also nrg (which is a Nero format) is similar to iso and we should consider it as iso.
When we support jpg then we can have jpeg too.
wav is an audio format and I guess it was added by mistake in "MOVIE_EXTENSIONS".