demux: add support for M2TS format recording (packetsize 192)
When the requested packet size is 192 bytes, set ts tap filter flag
0x80000000 to enable timestamps to be prepended to the ts packets.
This can be particularly useful in scrambled recordings (for pacing and
winding)

NOTE: this is not a linuxtv / dvb api standard flag, only a handfull of
dvb adapters currently implements it. But as there is no official way to
set the ts output format yet, we might as well use this flag for now.